requirements.txt
absl-py==2.1.0
alembic==1.13.2
aniso8601==9.0.1
attrs==23.2.0
blinker==1.8.2
cachetools==5.3.3
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==3.0.0
contourpy==1.2.1
cycler==0.12.1
Deprecated==1.2.14
docker==7.1.0
entrypoints==0.4
Flask==3.0.3
flatbuffers==24.3.25
fonttools==4.53.0
gitdb==4.0.11
GitPython==3.1.43
graphene==3.3
graphql-core==3.2.3
graphql-relay==3.2.0
greenlet==3.0.3
gunicorn==22.0.0
idna==3.7
importlib_metadata==7.1.0
itsdangerous==2.2.0
jax==0.4.30
jaxlib==0.4.30
Jinja2==3.1.4
joblib==1.4.2
kiwisolver==1.4.5
Mako==1.3.5
Markdown==3.6
MarkupSafe==2.1.5
matplotlib==3.9.0
mediapipe==0.10.14
ml-dtypes==0.4.0
mlflow==2.14.2
numpy==1.26.4
opencv-contrib-python==4.10.0.84
opencv-python==4.10.0.84
opentelemetry-api==1.25.0
opentelemetry-sdk==1.25.0
opentelemetry-semantic-conventions==0.46b0
opt-einsum==3.3.0
packaging==24.1
pandas==2.2.2
pillow==10.4.0
protobuf==4.25.3
psycopg2==2.9.9
psycopg2-binary==2.9.9
pyarrow==15.0.2
pycparser==2.22
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
querystring-parser==1.2.4
requests==2.32.3
scikit-learn==1.5.1
scipy==1.14.0
seaborn==0.13.2
setuptools==70.2.0
six==1.16.0
smmap==5.0.1
sounddevice==0.4.7
SQLAlchemy==2.0.31
SQLAlchemy-Utils==0.41.2
sqlparse==0.5.0
threadpoolctl==3.5.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
Werkzeug==3.0.3
wrapt==1.16.0
zipp==3.19.2
